server/pyproject.toml
[project]
name = "bratan-relay"
version = "0.1.0"
description = "Signaling relay for БРАТАН contacts. Auth via ed25519 challenge."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "pynacl>=1.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
